refactor(foundation): move SessionState+InteractionRecord + add clock hook (A7)
Phase 6 / Sub-phase A7: Moves the session tracking data types to foundation
while preserving time-travel integration via a pluggable clock.

Added mockforge-foundation::clock:
- `fn now() -> DateTime<Utc>` — default impl returns `Utc::now()`
- `fn set_clock(ClockFn)` — allows registering a custom clock (OnceLock)

Moved to mockforge-foundation::intelligent_behavior::session:
- InteractionRecord (pure data + builder methods)
- SessionState (uses clock::now() for created_at/last_activity)

Core's time_travel::register_global_clock() now also registers its `now` fn
with foundation so `SessionState::new()`/`touch()` honor virtual clocks.

Migrated mockforge-vbr/src/session.rs to use foundation's SessionState.

Core re-exports preserved — `mockforge_core::intelligent_behavior::types::SessionState`
still works for all existing consumers.

Workspace clean, 36 foundation tests passing.
